Add offset functionality to header levels
The `--offset` flag is useful in notebooks where the title of the notebook is a level 1 header while the rest of the section heads are of level 2. If the script is run on such a notebook without the `--offset` then the level 2 headers are added with a tab indent which breaks the structure of a markdown bullet list. If section heads start with level 2 then use `--offset 1`.

Also added code to replace non-alphanumeric characters in anchor id with `_` character. This was the cause of some broken anchor links.